+++ wsgidav/server/server_cli.py
@@ -131,10 +131,9 @@ See https://github.com/mar10/wsgidav for additional information.
         help="port to serve on (default: 8080)",
     )
     parser.add_argument(
-        "-H",
-        "--host",  # '-h' conflicts with --help
+        "-H",  # '-h' conflicts with --help
+        "--host",
         dest="host",
-        # default="localhost",
         help=(
             "host to serve from (default: localhost). 'localhost' is only "
             "accessible from the local computer. Use 0.0.0.0 to make your "
@@ -146,7 +145,6 @@ See https://github.com/mar10/wsgidav for additional information.
         "--root",
         dest="root_path",
         action=FullExpandedPath,
-        # type=arg_is_dir,
         help="path to a file system folder to publish as share '/'.",
     )
     parser.add_argument(
@@ -160,7 +158,14 @@ See https://github.com/mar10/wsgidav for additional information.
         choices=("builtin", "pyopenssl"),
         default="builtin",
         help="used by 'cheroot' server if SSL certificates are configured "
-        "(default: %(default)s.",
+        "(default: %(default)s).",
+    )
+    parser.add_argument(
+        "--auth",
+        choices=("anonymous", "nt", "pam-login"),
+        default=None,
+        help="Quick configuration of a default domain controller "
+        "(default: use settings from config file).",
     )
 
     qv_group = parser.add_mutually_exclusive_group()
@@ -196,16 +201,11 @@ See https://github.com/mar10/wsgidav for additional information.
     parser.add_argument(
         "-V",
         "--version",
-        # action="version",
-        # version=__version__,
         action="store_true",
         help="print version info and exit (may be combined with --verbose)",
     )
 
     args = parser.parse_args()
-
-    # if args.quiet and args.verbose > 3:
-    #     parser.error("-v and -q are mutually exclusive")
 
     args.verbose -= args.quiet
     del args.quiet
@@ -253,7 +253,7 @@ See https://github.com/mar10/wsgidav for additional information.
         print("Command line args:")
         for k, v in cmdLineOpts.items():
             print("    {:>12}: {}".format(k, v))
-    return cmdLineOpts
+    return cmdLineOpts, parser
 
 
 def _read_config_file(config_file, verbose):
@@ -306,18 +306,16 @@ def _read_config_file(config_file, verbose):
 
 def _init_config():
     """Setup configuration dictionary from default, command line and configuration file."""
-    cli_opts = _init_command_line_options()
+    cli_opts, parser = _init_command_line_options()
     cli_verbose = cli_opts["verbose"]
 
     # Set config defaults
-    # config = DEFAULT_CONFIG.copy()
     config = copy.deepcopy(DEFAULT_CONFIG)
 
     # Configuration file overrides defaults
     config_file = cli_opts.get("config_file")
     if config_file:
         file_opts = _read_config_file(config_file, cli_verbose)
-        # config.update(file_opts)
         util.deep_update(config, file_opts)
         if cli_verbose != DEFAULT_VERBOSE and "verbose" in file_opts:
             if cli_verbose >= 2:
@@ -336,15 +334,16 @@ def _init_config():
         config["port"] = cli_opts.get("port")
     if cli_opts.get("host"):
         config["host"] = cli_opts.get("host")
-    # if cli_opts.get("verbose") is not None:
-    if cli_opts.get("verbose") != DEFAULT_VERBOSE:
-        config["verbose"] = cli_opts.get("verbose")
     if cli_opts.get("profile") is not None:
         config["profile"] = True
     if cli_opts.get("server") is not None:
         config["server"] = cli_opts.get("server")
     if cli_opts.get("ssl_adapter") is not None:
         config["ssl_adapter"] = cli_opts.get("ssl_adapter")
+
+    # Command line overrides file only if -v or -q where passed:
+    if cli_opts.get("verbose") != DEFAULT_VERBOSE:
+        config["verbose"] = cli_opts.get("verbose")
 
     if cli_opts.get("root_path"):
         root_path = os.path.abspath(cli_opts.get("root_path"))
@@ -356,8 +355,59 @@ def _init_config():
         print("Configuration({}):\n{}".format(cli_opts["config_file"], pformat(config)))
 
     if not config["provider_mapping"]:
-        print("ERROR: No DAV provider defined. Try --help option.", file=sys.stderr)
-        sys.exit(-1)
+        parser.error("No DAV provider defined.")
+
+    # Quick-configuration of DomainController
+    auth = cli_opts.get("auth")
+    auth_conf = config.get("http_authenticator", {})
+    if auth and auth_conf.get("domain_controller"):
+        parser.error(
+            "--auth option can only be used when no domain_controller is configured"
+        )
+
+    if auth == "anonymous":
+        if config["simple_dc"]["user_mapping"]:
+            parser.error(
+                "--auth=anonymous can only be used when no user_mapping is configured"
+            )
+        auth_conf.update(
+            {
+                "domain_controller": "wsgidav.dc.simple_dc.SimpleDomainController",
+                "accept_basic": True,
+                "accept_digest": True,
+                "default_to_digest": True,
+            }
+        )
+        config["simple_dc"]["user_mapping"] = {"*": True}
+    elif auth == "nt":
+        if config.get("nt_dc"):
+            parser.error(
+                "--auth=nt can only be used when no nt_dc settings are configured"
+            )
+        auth_conf.update(
+            {
+                "domain_controller": "wsgidav.dc.nt_dc.NTDomainController",
+                "accept_basic": True,
+                "accept_digest": False,
+                "default_to_digest": False,
+            }
+        )
+        config["nt_dc"] = {}
+    elif auth == "pam-login":
+        if config.get("pam_dc"):
+            parser.error(
+                "--auth=pam-login can only be used when no pam_dc settings are configured"
+            )
+        auth_conf.update(
+            {
+                "domain_controller": "wsgidav.dc.pam_dc.PAMDomainController",
+                "accept_basic": True,
+                "accept_digest": False,
+                "default_to_digest": False,
+            }
+        )
+        config["pam_dc"] = {"service": "login"}
+    # print(config)
 
     if cli_opts.get("reload"):
         print("Installing paste.reloader.", file=sys.stderr)